Auto splitters UI/UX improvements
This commit changes multiple things regarding how auto splitters work and are handled by obs-livesplit-one

Instead of having to manually download the auto splitter files, obs-livesplit-one will now use the auto splitter list XML file and download the auto splitter files automatically.
Because of this, it can now display information on the auto splitter like the description and website.
The UI code might be a little jank but it should be updating properly depending on the user's actions.
The auto splitter list and files are also saved to a file in case there is no access to internet. They are saved in a new config folder that's gonna be used by obs-livesplit-one. However, while the list can be read from file, the auto splitters can't be activated from their files yet, this will need to be figured out at some point (probably when we will be able to save and retrieve auto splitter settings from the splits file)

To implement these functionalities, I relied on a few external crates to help with XML parsing and networking things, I assume some things could probably be implemented by livesplit-core but this will do for now.