Add UI to enable / disable auto splitter settings
This patch adds UI elements to be able to set the auto splitter custom settings .
The list of settings gets refreshed every time the user opens the properties window, you can then select a setting to change and you will be able to see it's tooltip, current value and enable or disable it.
There surely are things to improve in this but i think it's a good starting point.