Move obs_module_set_pointer back in lib.rs
The obs_module_set_pointer has been accidentally moved in the auto_splitter.rs file in a previous commit.
This meant that it wasn't being built when the auto splitting feature was turned off.
It turns out that it is actually essential for the obs module to register correctly, which means it wasn't working anymore when auto splitting was turned off.
This patch moves it back to the main lib.rs file so that it is exported no matter the configuration we're in.
I also added a comment to clarify what this is doing as it might not be very clear.

This fixes #41