Add advanced options to start a game
This patch adds the ability to provide command line arguments, a working
directory and environment variables to the game you want to start (it
also has a toggle to hide these settings if you don't need them). This
can be very helpful in case you need to start a game in a specific way.
For some reason, the 32 bits version of the Windows build doesn't seem
to work with the winsafe crate being used to parse the game arguments,
we'll have to see what we can do about that.