Fully show the auto splitter settings (#52)
I randomly came across the `obs_source_update_properties` function in
the documentation which allows us to dynamically rebuild the settings
GUI whenever necessary. This allows us to properly show all the auto
splitter settings and update them automatically when necessary.

Unfortunately the function is a little buggy in OBS and can cause
internal race conditions that crash OBS. It's probably rather rare that
this happens, so we can probably move forward with this implementation
regardless for now. There is as far as I can tell no way to work around
these issues.

The issue is tracked here:
https://github.com/obsproject/obs-studio/issues/10423